12/31/2013:21:13 - Paint tool functional
Selecting the paint tool, and then left clicking anywhere on the visible
portion of the map allows the user to replace the existing tile with the
currently selected tile on the sidebar. Does not occur if the mouse is
over the sidebar or top bar.

Need to draw icon of tool next to the drop down box and implement the
fill tool and cursor